+++ python/python/glide/async_commands/core.py
@@ -2424,6 +2424,61 @@ class CoreCommands(Protocol):
         )
         return cast(bool, await self._execute_command(RequestType.PExpireAt, args))
 
+    async def expiretime(self, key: str) -> int:
+        """
+        Returns the absolute Unix timestamp (since January 1, 1970) at which
+        the given `key` will expire, in seconds.
+        To get the expiration with millisecond precision, use `pexpiretime`.
+
+        See https://valkey.io/commands/expiretime/ for details.
+
+        Args:
+            key (str): The `key` to determine the expiration value of.
+
+        Returns:
+            int: The expiration Unix timestamp in seconds, -2 if `key` does not exist or -1 if `key` exists but has no associated expire.
+
+        Examples:
+            >>> await client.expiretime("my_key")
+                -2 # 'my_key' doesn't exist.
+            >>> await client.set("my_key", "value")
+            >>> await client.expiretime("my_key")
+                -1 # 'my_key' has no associate expiration.
+            >>> await client.expire("my_key", 60)
+            >>> await client.expiretime("my_key")
+                1718614954
+
+        Since: Redis version 7.0.0.
+        """
+        return cast(int, await self._execute_command(RequestType.ExpireTime, [key]))
+
+    async def pexpiretime(self, key: str) -> int:
+        """
+        Returns the absolute Unix timestamp (since January 1, 1970) at which
+        the given `key` will expire, in milliseconds.
+
+        See https://valkey.io/commands/pexpiretime/ for details.
+
+        Args:
+            key (str): The `key` to determine the expiration value of.
+
+        Returns:
+            int: The expiration Unix timestamp in milliseconds, -2 if `key` does not exist, or -1 if `key` exists but has no associated expiration.
+
+        Examples:
+            >>> await client.pexpiretime("my_key")
+                -2 # 'my_key' doesn't exist.
+            >>> await client.set("my_key", "value")
+            >>> await client.pexpiretime("my_key")
+                -1 # 'my_key' has no associate expiration.
+            >>> await client.expire("my_key", 60)
+            >>> await client.pexpiretime("my_key")
+                1718615446670
+
+        Since: Redis version 7.0.0.
+        """
+        return cast(int, await self._execute_command(RequestType.PExpireTime, [key]))
+
     async def ttl(self, key: str) -> int:
         """
         Returns the remaining time to live of `key` that has a timeout.